+++ mutatest/__init__.py
@@ -1,6 +1,6 @@
 """Mutation initialization.
 """
-__version__ = "3.0.2"
+__version__ = "3.1.0"
 
 __title__ = "mutatest"
 __description__ = "Python mutation testing: test your tests!"